+++ src/capacitor-run.ts
@@ -4,6 +4,8 @@ import { exists } from './analyzer';
 import { CapacitorPlatform } from './capacitor-platform';
 import { InternalCommand } from './command-name';
 import { getOutputChannel } from './extension';
+import { ionicBuild } from './ionic-build';
+import { ionicState } from './ionic-tree-provider';
 import { MonoRepoType } from './monorepo';
 import { Project } from './project';
 
@@ -16,11 +18,20 @@ import { Project } from './project';
  * @returns string
  */
 export function capacitorRun(project: Project, platform: CapacitorPlatform): string {
+    let preop = '';
+
+	// If the user modified something in the editor then its likely they need to rebuild the app before running
+	if (ionicState.projectDirty) {
+		const channel = getOutputChannel();
+		channel.appendLine('[Ionic] Rebuilding as you changed your project...');
+		preop = ionicBuild(project) + ' && ';		
+	}
+
 	switch (project.repoType) {
 		case MonoRepoType.none:
 		case MonoRepoType.folder:
-		case MonoRepoType.npm: return capRun(platform, project.repoType);
-		case MonoRepoType.nx: return nxRun(project, platform);
+		case MonoRepoType.npm: return preop + capRun(platform, project.repoType);
+		case MonoRepoType.nx: return preop + nxRun(project, platform);
 		default: throw new Error('Unsupported Monorepo type');
 	}
 }
@@ -34,6 +45,7 @@ function capRun(platform: CapacitorPlatform, repoType: MonoRepoType): string {
 	const liveReload = vscode.workspace.getConfiguration('ionic').get('liveReload');
 	const externalIP = vscode.workspace.getConfiguration('ionic').get('externalAddress');
 	let capRunFlags = liveReload ? ' -l' : '';
+
 
 	if (exists('@ionic-enterprise/auth') && liveReload) {
 		capRunFlags = '';