+++ src/command-name.ts
@@ -10,6 +10,7 @@ export enum CommandName {
 	Login = 'ionic.login',
 	Stop = 'ionic.stop',
 	Rebuild = 'ionic.rebuild',
+	Function = 'ionic.function',
 	Open = 'ionic.open',
 	SkipLogin = 'ionic.skipLogin',
 	Upgrade = 'ionic.upgrade',
@@ -18,6 +19,7 @@ export enum CommandName {
 	BuildConfig = 'ionic.buildConfig',
 	DebugMode = 'ionic.debugMode',
 	RunMode = 'ionic.runMode',
+	SelectDevice = 'ionic.selectDevice'
 }
 
 export enum InternalCommand {