+++ src/recommend.ts
@@ -22,6 +22,9 @@ import { CapacitorPlatform } from './capacitor-platform';
 import { addScripts } from './scripts';
 import { Context } from './context-variables';
 import { ionicState } from './ionic-tree-provider';
+import { getAndroidWebViewList } from './android-debug-list';
+import { getDebugBrowserName } from './editor-preview';
+import { features } from './features';
 
 export async function getRecommendations(project: Project, context: vscode.ExtensionContext, packages: any): Promise<void> {
 	if (isCapacitor() && !isCordova()) {
@@ -29,8 +32,9 @@ export async function getRecommendations(project: Project, context: vscode.Exten
 
 		const hasCapIos = project.hasCapacitorProject(CapacitorPlatform.ios);
 		const hasCapAndroid = project.hasCapacitorProject(CapacitorPlatform.android);
-		const title = (ionicState.debugMode) ? 'Debug On Web' : 'Run On Web';
-		const type = (ionicState.debugMode) ? TipType.Debug : TipType.Run;
+		const title = (ionicState.webDebugMode) ? 'Debug On Web' : 'Run On Web';
+		const tooltip = (ionicState.webDebugMode) ? `Debug using ${getDebugBrowserName()}. The browser can be changed in Settings.` : 'Run a developement server and open using the default web browser';
+		const type = (ionicState.webDebugMode) ? TipType.Debug : TipType.Run;
 		project.add(
 			new Tip(
 				title, '', type, 'Serve', undefined, 'Running on Web', `Project Served`)
@@ -43,9 +47,11 @@ export async function getRecommendations(project: Project, context: vscode.Exten
 				.canStop()
 				.contextIf(Context.debugMode, false)
 				.canAnimate()
+				.setTooltip(tooltip)
 		);
 		// project.add(new Tip('View In Editor', '', TipType.Run, 'Serve', undefined, 'Running on Web', `Project Served`).setAction(viewInEditor, 'http://localhost:8100'));
 		const runPoints = [
+			{ text: 'Copying web assets', title: 'Copying...' },
 			{ text: 'ng run app:build', title: 'Building Web...' },
 			{ text: 'capacitor run', title: 'Syncing...' },
 			{ text: '✔ update ios', title: 'Building Native...' },
@@ -56,23 +62,27 @@ export async function getRecommendations(project: Project, context: vscode.Exten
 		];
 
 		if (hasCapAndroid) {
-			project.add(new Tip('Run On Android', '', TipType.Run, 'Run', undefined, 'Running', 'Project is running')
+			const title = (ionicState.selectedAndroidDevice) ? `Run on ${ionicState.selectedAndroidDeviceName}` : 'Run on Android';
+			project.add(new Tip(title, '', TipType.Run, 'Run', undefined, 'Running', 'Project is running')
 				.showProgressDialog()
 				.requestDeviceSelection()
 				.setDynamicCommand(capacitorRun, project, CapacitorPlatform.android)
 				.setSecondCommand('Getting Devices', capacitorDevicesCommand(CapacitorPlatform.android))
 				.setData(project.projectFolder())
 				.setRunPoints(runPoints)
+				.setContextValue(Context.selectDevice)
 			);
 		}
 		if (hasCapIos) {
-			project.add(new Tip('Run On iOS', '', TipType.Run, 'Run', undefined, 'Running', 'Project is running')
+			const title = (ionicState.selectedIOSDevice) ? `Run on ${ionicState.selectedIOSDeviceName}` : 'Run on iOS';
+			project.add(new Tip(title, '', TipType.Run, 'Run', undefined, 'Running', 'Project is running')
 				.showProgressDialog()
 				.requestDeviceSelection()
 				.setDynamicCommand(capacitorRun, project, CapacitorPlatform.ios)
 				.setSecondCommand('Getting Devices', capacitorDevicesCommand(CapacitorPlatform.ios))
 				.setData(project.projectFolder())
 				.setRunPoints(runPoints)
+				.setContextValue(Context.selectDevice)
 			);
 		}
 
@@ -99,6 +109,13 @@ export async function getRecommendations(project: Project, context: vscode.Exten
 				.showProgressDialog()
 				.setDynamicCommand(capacitorOpen, project, CapacitorPlatform.android)
 			);
+		}
+		
+		if (features.debugAndroid) { // Experimental Feature
+			const r = project.setGroup(`Debug`, 'Running Ionic applications you can debug', TipType.Debug, false);
+			r.whenExpanded = async () => {
+				return getAndroidWebViewList(hasCapAndroid);
+			};
 		}
 	}