+++ src/rules-capacitor.ts
@@ -63,6 +63,15 @@ export function checkCapacitorRules(project: Project) {
 		// TODO: Complete work
 		// checkAndroidManifest();
 	}
+
+	// Ionic CLI unlock live reload
+	if (!exists('@ionic/cli') && isIonicBasedProject()) {
+		project.recommendAdd('@ionic/cli', '@ionic/cli',
+			'Add the Ionic CLI for Live Reload',
+			'The Ionic CLI adds the "Live Reload" feature to quickly refresh your application on a device after a code change.',
+			true);
+	}
+
 	if (exists('cordova-plugin-x-socialsharing')) {
 		// TODO: Verify that Android Manifest contains
 		// <uses-permission android:name="android.permission.QUERY_ALL_PACKAGES" />
@@ -87,28 +96,38 @@ export function capacitorRecommendations(project: Project): Tip[] {
 		tips.push(new Tip(
 			'Add Capacitor Integration', '', TipType.Capacitor, 'Add the Capacitor integration to this project',
 			[
-				npmInstall('@capacitor/core@latest', '--save','-E'),
+				npmInstall('@capacitor/core@latest', '--save', '-E'),
 				npmInstall('@capacitor/cli@latest', '-D', '-E'),
 				npmInstall(`@capacitor/app @capacitor/core @capacitor/haptics @capacitor/keyboard @capacitor/status-bar`),
 				`${local}npx capacitor init "${project.name}" "${asAppId(project.name)}" --web-dir www`
 			],
 			'Add Capacitor', 'Capacitor added to this project',
-			'https://capacitorjs.com/docs/cordova/migrating-from-cordova-to-capacitor'
+			'https://capacitorjs.com/docs/cordova/migrating-from-cordova-to-capacitor',
+			'Adding Capacitor to the project...'
 		).showProgressDialog());
 	} else {
 		if (!project.hasCapacitorProject(CapacitorPlatform.android)) {
 			tips.push(new Tip(
-				'Add Android Project', '', TipType.Capacitor, 'Add Android support to your Capacitor project',
-				[npmInstall('@capacitor/android'), capacitorAdd(project, CapacitorPlatform.android)], 'Add Android', 'Android support added to your project').showProgressDialog());
+				'Add Android Project', '', TipType.Capacitor, 'Add Android support to your Capacitor project?',
+				[npmInstall('@capacitor/android'), capacitorAdd(project, CapacitorPlatform.android)],
+				'Add Android',
+				'Android support added to your project', undefined,
+				'Adding Native Android Project...')
+				.showProgressDialog()
+			);
 		}
 
 		if (!project.hasCapacitorProject(CapacitorPlatform.ios)) {
 			tips.push(new Tip(
-				'Add iOS Project', '', TipType.Capacitor, 'Add iOS support to your Capacitor project',
-				[npmInstall('@capacitor/ios'), capacitorAdd(project, CapacitorPlatform.ios)], 'Add iOS', 'iOS support added to your project').showProgressDialog());
+				'Add iOS Project', '', TipType.Capacitor, 'Add iOS support to your Capacitor project?',
+				[npmInstall('@capacitor/ios'), capacitorAdd(project, CapacitorPlatform.ios)],
+				'Add iOS',
+				'iOS support added to your project', undefined,
+				'Adding Native iOS Project...')
+				.showProgressDialog()
+			);
 		}
 	}
-
 
 	// List of incompatible plugins
 	tips.push(incompatiblePlugin('cordova-plugin-admobpro', 'https://github.com/ionic-team/capacitor/issues/1101'));
@@ -179,6 +198,10 @@ export function capacitorRecommendations(project: Project): Tip[] {
 	tips.push(replacementPlugin('cordova-plugin-statusbar', '@capacitor/status-bar', 'https://capacitorjs.com/docs/apis/status-bar'));
 	tips.push(replacementPlugin('phonegap-plugin-push', '@capacitor/push-notifications', 'https://capacitorjs.com/docs/apis/push-notifications'));
 	return tips;
+}
+
+function isIonicBasedProject() {
+	return (exists('@ionic/angular') || exists('@ionic/react') || exists('@ionic/vue'));
 }
 
 // Capacity Android 3.2.3 added proguard rules for Capacitor for release build