+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## Changelog
+### Version 1.4.2
+- When re-running on iOS or Android the device does not need to be reselected
+- For projects not integrated with Ionic automate initialization of custom project with ionic:build and ionic:serve scripts
+- If you change your project and click run on iOS/Android it will rebuild the app before running
+- Undisplayable Ansi color text is removed from the output window
+- Recommendation for Ionic CLI installation to allow Live Reload to work for Ionic projects
+
 ### Version 1.4.1
 - Add folder style mono repos
 - Option to debug for web browser